services/user_service/requirements.txt
fastapi
uvicorn
pyjwt
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt